requirements.txt
streamlit>=1.41.1
google-generativeai>=0.8.4
python-dotenv>=1.0.1
Pillow>=11.1.0
pandas>=2.2.3
openpyxl>=3.1.2  # For Excel export support 